services/streamlit/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.29.0